META-198: Added an object metadata, a track info, and the best video frame to an analytics action
META-206: New analytics objects stop to appear in Right Panel
META-208: Make right panel tiles request screenshot preview after a configurable timeout
META-212: Added an analytics logging directory setting to analytics_logging.ini
META-212: Added separate analytics metadata loggers for primary and secondary streams
META-212: Create metadata logger in WidgetAnalyticsController only if logging is enabled in the ini
META-212: Draft implementation of frame/metadata logging on the client side
META-212: Draft implementation of frame/metadata logging on the server side.
META-212: Fixed analytics_logging.ini setting comment
META-212: Fixed unit test placeholder_binder_ut
META-212: Improve metadata logging format
META-81: Make analytics object always fit into the parent camera widget
VMS-13510: DW Edge localhost workaround
VMS-13595: Fix coding style and add/improve comments
VMS-13595: Live Cache: Moved min/max size to mediaserver.ini and changed defaults to 67..100 s
VMS-13595: mediaserver.ini forceLiveCacheForPrimaryStream transformed into a System global setting
VMS-13673: Don't return old data from GopKeeper if camera stream is not running

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/server/plugins/analytics/stub_analytics_plugin/src/nx/vms_server_plugins/analytics/stub/device_agent.cpp
@@ -299,7 +299,7 @@ static IObjectMetadata* makeObjectMetadata(
     objectMetadata->setTypeId(objectTypeId);
     objectMetadata->setId(objectId);
     objectMetadata->setBoundingBox(IObjectMetadata::Rect((float) offset,
-        (float) offset + 0.05 * objectIndex, 0.25F, 0.25F));
+        (float) offset + 0.05F * (float) objectIndex, 0.25F, 0.25F));
 
     if (generatePreviewAttributes)
     {
